p_19_CHAVEZ_SCHOOL/myapp/requirements.txt
alabaster==0.7.13
alembic==1.12.0
appdirs==1.4.4
astroid==1.6.6
attrs==23.1.0
autoflake==0.6.6
autopep8==1.7.0
Babel==2.12.1
bandit==1.6.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmakelint==1.4.2
coala==0.11.0
coala-bears==0.11.1
coala-utils==0.6.7
colorama==0.3.9
colorlog==2.10.0
cppclean==0.12
cpplint==1.6.1
cryptography==41.0.4
dennis==0.9
dependency-management==0.4.0
dill==0.3.7
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
email-validator==2.0.0.post2
eradicate==0.1.6
Faker==19.6.1
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
guess-language-spirit==0.5.3
html-linter==0.3.0
idna==3.4
imagesize==1.4.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jwt==1.3.1
lazy-object-proxy==1.9.0
libclang-py3==3.4.0
Mako==1.2.4
mando==0.3.3
MarkupSafe==2.1.3
mccabe==0.7.0
munkres3==1.0.5.5
mypy-lang==0.4.6
nbformat==4.4.0
nltk==3.8.1
packaging==23.1
pbr==5.11.1
platformdirs==3.10.0
polib==1.2.0
proselint==0.7.0
psutil==5.9.5
psycopg2==2.9.7
py-spy==0.3.14
pycodestyle==2.11.0
pycparser==2.21
pydeps==1.12.17
pydocstyle==1.1.1
pyflakes==1.4.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==1.9.4
PyPrint==0.2.6
pyroma==2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==3.13
radon==1.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
restructuredtext-lint==1.0.1
rpds-py==0.10.3
rstcheck==2.2
safety==0.5.1
sarge==0.1.7.post1
scspell3k==2.2
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==1.4.9
SQLAlchemy==2.0.20
stdlib-list==0.9.0
stevedore==5.1.0
template-remover==0.1.9
termcolor==1.1.0
toml==0.10.2
tomlkit==0.12.1
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.7.1
unidiff==0.5.5
urllib3==2.0.5
vprof==0.38
vulture==0.10
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
WTForms-Ext==0.5
WTForms-SQLAlchemy==0.3
yamllint==1.6.1
yapf==0.14.0
